Add admin route to get all invite requests

+++ routes/admin.php
@@ -1,6 +1,5 @@
 <?php
 
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -42,9 +41,11 @@ Route::get('/profiles', [\App\Http\Controllers\Admin\ProfileController::class, '
 
 Route::group(['prefix' => 'invites'], function () {
     // Get all invites
-    Route::get('/', [\App\Http\Controllers\InviteController::class, 'all']);
+    Route::get('/', [\App\Http\Controllers\Admin\InviteController::class, 'all']);
+    // Get all invite requests
+    Route::get('/requests', [\App\Http\Controllers\Admin\InviteRequestController::class, 'all']);
     // Create new invite(s)
-    Route::get('/create', [\App\Http\Controllers\InviteController::class, 'create']);
+    Route::get('/create', [\App\Http\Controllers\Admin\InviteController::class, 'create']);
     // Delete invite
-    Route::delete('/{email}', [\App\Http\Controllers\InviteController::class, 'delete']);
+    Route::delete('/{email}', [\App\Http\Controllers\Admin\InviteController::class, 'delete']);
 });
